replaced pow() with lpow()
pow() function is inaccurate.  Was using to raise base of 10 to the power of the range digit in the first byte of the packet.  E.g. 1000000 was coming b back as 999999.98 instead of 1000000.00.   Since both the base and range are integers, I made an integer based fn that returns a long integer.